refactor: organiza código em classes separadas para conexão e lógica
