+++ Controle-Ponto/src/main/java/unifacs/a3/App.java
@@ -32,7 +32,7 @@ public class App {
                 int senha_user = scan.nextInt();
 
                 // Verifica se o usuário existe no BD
-                Resultado r1 = verificaUser(nome_user, senha_user);
+                Resultado r1 = Usuario.verificaUser(nome_user, senha_user);
          if (r1.user_encontrado()) {
         // Menu inicial
         System.out.println("Selecione uma opção:\n1-Registrar ponto\n2-Cadastrar Funcionario\n3-Registrar justificativa de atraso\n5-verificar atrasos");
@@ -49,7 +49,7 @@ public class App {
                     // ----------------- REGISTRA ENTRADA -----------------
                     if (resposta == 1) {
                         LocalDateTime entrada = LocalDateTime.now();
-                        boolean registrado = registra_Hora(entrada, r1.id_user());
+                        boolean registrado =Usuario.registra_Hora(entrada, r1.id_user());
 
                         if (registrado) {
                             System.out.println("Sua entrada foi registrada com sucesso");
@@ -57,27 +57,8 @@ public class App {
 
                     // ----------------- REGISTRA SAÍDA -----------------
                     } else {
-                        Connection connect = conexao_BD();
-                        Statement stmt = connect.createStatement();
-
-                        // Busca o último registro sem saída
-                        ResultSet rs = stmt.executeQuery(
-                            "SELECT id FROM horarios WHERE saida IS NULL ORDER BY id DESC LIMIT 1"
-                        );
-
-                        if (rs.next()) {
-                            int id_buscado = rs.getInt("id");
-
-                            // Atualiza a saída no registro encontrado
-                            PreparedStatement ps = connect.prepareStatement(
-                                "UPDATE horarios SET saida = ? WHERE id = ?"
-                            );
-                            ps.setObject(1, Timestamp.valueOf(LocalDateTime.now()));
-                            ps.setInt(2, id_buscado);
-                            ps.executeUpdate();
-
-                            System.out.println("Saída registrada com sucesso");
-                        }
+                        Usuario.registra_Saida( );
+                        System.out.println("Sua saída foi registrada com sucesso");
                     }
                 
                     break;
@@ -90,15 +71,7 @@ public class App {
 
                 System.out.println("Crie uma nova senha:");
                 int senha = scan.nextInt();
-
-                try (Connection conectando = conexao_BD()) {
-                    PreparedStatement ps = conectando.prepareStatement(
-                        "INSERT INTO usuario (nome, senha) VALUES (?, ?)"
-                    );
-                    ps.setString(1, nome);
-                    ps.setInt(2, senha);
-                    ps.executeUpdate();
-                }
+                Usuario.cadastra_Usuario(nome, senha);
                 break;
 
             // ------------------------- CASE 3: Justificar atraso -------------------------
@@ -113,7 +86,9 @@ public class App {
 
                break;
 
-               case 5:
+                 // ------------------------- CASE 5: Verificar Atrasos -------------------------
+
+            case 5:
                System.out.println("Verificando atrasos...");
                verifica_atraso(r1.id_user());
             break;
@@ -128,7 +103,7 @@ public class App {
             alternativa=scan.nextInt();
             if(alternativa==1){
                 System.out.println("digite seu email:");
-                recupera_senha();
+                Usuario.recupera_senha();
 
             }
         
@@ -138,84 +113,16 @@ public class App {
 
     // ------------------------- MÉTODOS AUXILIARES -------------------------
 
-    // Faz a conexão com o banco de dados PostgreSQL
-    public static Connection conexao_BD() {
-        String url = "jdbc:postgresql://aws-1-us-east-1.pooler.supabase.com:6543/postgres?sslmode=require";
-        String user = "postgres.iczuamsiqqdvifexshbf";
-        String password = "miniProject25@";
-        Connection conn = null;
+    
 
-        try {
-            conn = DriverManager.getConnection(url, user, password);
-            System.out.println("Conexão bem-sucedida com o PostgreSQL!");
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        return conn;
-    }
 
-    // Verifica se o usuário existe no banco
-    public static Resultado verificaUser(String nome_user, int senha_user) throws Exception {
-        Connection connect = conexao_BD();
-        boolean user_cadastrado = false;
-        int id = 0;
 
-        String sql = "SELECT id FROM usuario WHERE nome = ? AND senha = ?";
-        PreparedStatement ps = connect.prepareStatement(sql);
-        ps.setString(1, nome_user);
-        ps.setInt(2, senha_user);
+    //Justificativa de atraso-Elísio
 
-        ResultSet rs = ps.executeQuery();
-
-        // Se encontrou uma linha, usuário existe
-        if (rs.next()) {
-            user_cadastrado = true;
-            id = rs.getInt("id");
-        
-        }
-
-        return new Resultado(id, user_cadastrado);
-    }
-
-    // Registra a hora de entrada no banco
-    public static boolean registra_Hora(LocalDateTime entrada, int usuario_id) throws Exception {
-        Connection connect = conexao_BD();
-        boolean horario_registrado = false;
-
-        PreparedStatement ps = connect.prepareStatement(
-            "INSERT INTO horarios (usuario_id, entrada) VALUES (?, ?)"
-        );
-        ps.setInt(1, usuario_id);
-        ps.setObject(2, entrada);
-        ps.executeUpdate();
-
-        return horario_registrado = true;
-    }
-    public static void recupera_senha() throws Exception{
-        Scanner scan=new Scanner(System.in);
-        String email= scan.nextLine();
-        Connection con= conexao_BD();
-       String sql = "SELECT  FROM usuario WHERE email = ?";
-        PreparedStatement ps = con.prepareStatement(sql);
-        ps.setString(1, email);
-        ResultSet rs=ps.executeQuery();
-        if(rs.next()){
-            System.out.println("Digite sua nova senha:");
-            int senha_nova=scan.nextInt();
-            ps=con.prepareStatement("UPDATE usuario set senha=? where email=?");
-            ps.setInt(1,senha_nova);
-            ps.setString(2,email);
-            ps.executeUpdate();
-             System.out.println("Senha alterada com sucesso!");
-        }else{
-            System.out.println("Usuário não encontrado!");
-        
-        }
-    }
     public static void justifica_atraso(int id) throws Exception{
         Scanner scan=new Scanner(System.in);
                 int justificativa = scan.nextInt();
-            Connection connect=conexao_BD();
+            Connection connect=Conexao.conexao_BD();
                 switch (justificativa) {
                     case 1:
                       PreparedStatement pr=connect.prepareStatement("INSERT INTO justificativa(tipo,usuario_id) values(?,?)");
@@ -257,7 +164,7 @@ public class App {
 
 
     public static void verifica_atraso(int id) throws Exception {
-    Connection connect = conexao_BD();
+    Connection connect = Conexao.conexao_BD();
     Statement stmt = connect.createStatement();
 
     // Definindo horário limite de entrada (08:00 da manhã)
